Code cleanup and UI improvements
- Remove BarChart3 icon from app header
- Remove duplicate utils file (shared/lib/utils.ts)
- Consolidate all utility functions into lib/utils.ts
- Update all imports to use @/lib/utils
- Delete backup file (member-details.backup.tsx)
- Replace large dashboard stat cards with compact header badges
- Show total members and biggest chapter in header badges
- Reduce visual clutter on dashboard
- Fix import paths for better maintainability
