+++ frontend/src/app/App.tsx
@@ -7,7 +7,7 @@ import {
   useLocation,
 } from "react-router-dom";
 import { QueryClientProvider } from '@tanstack/react-query';
-import { BarChart3, Settings, Home } from "lucide-react";
+import { Settings, Home } from "lucide-react";
 import { Button } from '@/components/ui/button';
 
 // Import components
@@ -74,7 +74,6 @@ function AppContent() {
         <header className="sticky top-0 z-40 border-b bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60">
           <div className="flex h-16 items-center justify-between px-6">
             <div className="flex items-center gap-3">
-              <BarChart3 className="h-7 w-7 text-primary" />
               <div className="flex flex-col">
                 <h1 className="text-xl font-semibold text-foreground">
                   📊 BNI PALMS Analysis
